# doc/api.yaml
openapi: 3.0.0
paths:
...
  /artist:
    get:
      description: Gets all artists
      operationId: ArtistController_findAll
      parameters: []
      responses:
        '200':
          description: Returns all artists records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Artist'
      summary: Get all artists
      tags:
        - Artist
    post:
      description: Add new artist
      operationId: ArtistController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArtistDto'
      responses:
        '201':
          description: Returns if artist has been created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Returns if request body does not contain required fields
      summary: Add new artist
      tags:
        - Artist
  /artist/{id}:
    get:
      description: Get single artist by id
      operationId: ArtistController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Artist ID
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Returns if artist record exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Returns if artist ID is not valid UUID
        '404':
          description: Returns if artist record does't exist
      summary: Get single artist by id
      tags:
        - Artist
    put:
      description: Update artist information by UUID
      operationId: ArtistController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Artist ID
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtistDto'
      responses:
        '200':
          description: Returns if artist has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Returns if artist ID is not valid UUID
        '404':
          description: Returns if artist does not exist
      summary: Update artist information
      tags:
        - Artist
    delete:
      description: Delete artist from library
      operationId: ArtistController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: Artist ID
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Returns if artist has been removed
        '400':
          description: Returns if artist ID is not valid UUID
        '404':
          description: Returns if artist does not exist
      summary: Delete artist
      tags:
        - Artist
...
  /album:
    get:
      description: Gets all library albums list
      operationId: AlbumController_findAll
      parameters: []
      responses:
        '200':
          description: Returns all albums records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Album'
      summary: Get albums list
      tags:
        - Album
    post:
      description: Add new album information
      operationId: AlbumController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlbumDto'
      responses:
        '201':
          description: Returns if album has been created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Returns if request body does not contain required fields
      summary: Add new album
      tags:
        - Album
  /album/{id}:
    get:
      description: Get single album by id
      operationId: AlbumController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Album ID
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Returns if album record exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Returns if album ID is not valid UUID
        '404':
          description: Returns if album record does't exist
      summary: Get single album by id
      tags:
        - Album
    put:
      description: Update library album information by UUID
      operationId: AlbumController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Album ID
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlbumDto'
      responses:
        '200':
          description: Returns if album has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Returns if album ID is not valid UUID
        '404':
          description: Returns if album does not exist
      summary: Update album information
      tags:
        - Album
    delete:
      description: Delete album from library
      operationId: AlbumController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: Album ID
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Returns if album has been removed
        '400':
          description: Returns if album ID is not valid UUID
        '404':
          description: Returns if album does not exist
      summary: Delete album
      tags:
        - Album
...
  /favs/album/{id}:
    post:
      description: Add album to the favorites
      operationId: FavsController_addAlbum
      parameters:
        - name: id
          required: true
          in: path
          description: Album ID
          schema:
            format: uuid
            type: string
      responses:
        '201':
          description: Returns if album has been added to the favorites
        '400':
          description: Returns if album ID is not valid UUID
        '422':
          description: Returns if album with ID does not exist
      summary: Add album to the favorites
      tags:
        - Favs
    delete:
      description: Delete album from favorites
      operationId: FavsController_removeAlbum
      parameters:
        - name: id
          required: true
          in: path
          description: Album ID
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Returns if album has been removed
        '400':
          description: Returns if album ID is not valid UUID
        '404':
          description: Returns if album is not favorite
      summary: Delete album from favorites
      tags:
        - Favs
  /favs/artist/{id}:
    post:
      description: Add artist to the favorites
      operationId: FavsController_addArtist
      parameters:
        - name: id
          required: true
          in: path
          description: Artist ID
          schema:
            format: uuid
            type: string
      responses:
        '201':
          description: Returns if artist has been added to the favorites
        '400':
          description: Returns if artist ID is not valid UUID
        '422':
          description: Returns if artist with ID does not exist
      summary: Add artist to the favorites
      tags:
        - Favs
    delete:
      description: Delete artist from favorites
      operationId: FavsController_removeArtist
      parameters:
        - name: id
          required: true
          in: path
          description: Artist ID
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Returns if artist has been removed
        '400':
          description: Returns if artist ID is not valid UUID
        '404':
          description: Returns if artist is not favorite
      summary: Delete artist from favorites
      tags:
        - Favs
info:
  title: Home Library Service
  description: Home music library service
  version: '1.0'
  contact: {}
...
servers: []
components:
  schemas:
...
    Artist:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Freddie Mercury
        grammy:
          type: boolean
          example: false
      required:
        - id
        - name
        - grammy
    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
        grammy:
          type: boolean
      required:
        - name
        - grammy
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
        grammy:
          type: boolean
...
    Album:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Innuendo
        year:
          type: integer
          example: 1991
        artistId:
          type: string
          description: Refers to Artist
          nullable: true
          format: uuid
      required:
        - id
        - name
        - year
        - artistId
    CreateAlbumDto:
      type: object
      properties:
        name:
          type: string
        year:
          type: integer
        artistId:
          type: string
          description: Refers to Artist
          nullable: true
          format: uuid
      required:
        - name
        - year
        - artistId
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
        year:
          type: integer
        artistId:
          type: string
          description: Refers to Artist
          nullable: true
          format: uuid
...